novos ajustes em testes.

+++ socket/src/anotacoes/Arquivo.java
@@ -7,9 +7,9 @@ public class Arquivo {
 	String raiz;
 	String origem;
 	String destino;
-	int tamanho;
+	long tamanho;
 	
-	public Arquivo(String raiz, String origem, String destino, int tamanho) {
+	public Arquivo(String raiz, String origem, String destino, long tamanho) {
 		super();
 		this.raiz = raiz;
 		this.origem = origem;
